Issue 40 code splitting started. In this commit, requirements.py, install_python_packages.py, global_variables.py have been created in the subfolder utils, and SlicerCARTConfigurationInitialWindow.py has been created, which represent the class SlicerCARTConfigurationInitialWindow in SlicerCART.py. The apps works the same as the actual main branch at the time of the commit. This commit shows concretely that it is possible to split the code in the slicer module. Thanks to Kuan for its initiation to this aspect. Next steps of code splitting to be discussed at the next Neuropoly Slicer meeting. Note that code here has been moved, not modified (except for few minors aspects related to importation). Thanks!
