+++ SlicerCART/src/SlicerCART.py
@@ -1,235 +1,142 @@
-# To install a package in slicer python environment, use the following command:
-# pip install --user package_name
-import os
-import logging
-import slicer
-import qt
-from slicer.ScriptedLoadableModule import *
-from slicer.util import VTKObservationMixin
-from glob import glob
-import re
-import time
-from pathlib import Path
-from threading import RLock
-from datetime import datetime
-import filecmp
-import shutil
-import numpy as np
-import vtk
-import random
-import colorsys
-import sys
-from functools import partial
-import copy
-
-
-# TODO: There is probably a more elegant way to install pacakages through the extension manager when the user installs the extension.
-# TODO: check if the package installed with error
-
-# Dictionary of required python packages and their import names
-REQUIRED_PYTHON_PACKAGES = {
-    "nibabel": "nibabel",
-    "pandas": "pandas",
-    "PyYAML": "yaml",
-    "pynrrd": "nrrd",
-    "slicerio": "slicerio",
-    "bids_validator": "bids_validator"
-}
-
-def check_and_install_python_packages():
-    missing_packages = []
-    
-    for pip_name, import_name in REQUIRED_PYTHON_PACKAGES.items():
-        try:
-            __import__(import_name)
-        except ImportError:
-            missing_packages.append(pip_name)
-
-    if missing_packages:
-        msg = "SlicerCART module: The following required python packages are missing:"
-        msg += "\n" + "\n".join(missing_packages)
-        msg += "\nWould you like to install them now?"
-        response = qt.QMessageBox.question(slicer.util.mainWindow(), 'Install Extensions', msg,
-                                           qt.QMessageBox.Yes | qt.QMessageBox.No)
-        if response == qt.QMessageBox.Yes:
-            for pip_name in missing_packages:
-                slicer.util.pip_install(pip_name)
-                # Wait for the installation to complete
-                slicer.app.processEvents()
-            qt.QMessageBox.information(slicer.util.mainWindow(), 'Restart Slicer',
-                                       'Slicer will now restart to complete the installation.')
-            slicer.app.restart()
-        else:
-            qt.QMessageBox.warning(slicer.util.mainWindow(), 'Missing Extensions',
-                                   'The SlicerCART module cannot be loaded without the required extensions.')
-
-
+from utils.requirements import *
+from utils.install_python_packages import *
 check_and_install_python_packages()
-
-from bids_validator import BIDSValidator
-import nibabel as nib
-import nrrd
-import pandas as pd
-import slicerio
-import yaml
-
-INPUT_FILE_EXTENSION = '*.nii.gz'
-INTERPOLATE_VALUE = 0
-
-DEFAULT_VOLUMES_DIRECTORY = ''
-DEFAULT_SEGMENTATION_DIRECTORY = ''
-
-REQUIRE_VOLUME_DATA_HIERARCHY_BIDS_FORMAT = False
-
-IS_CLASSIFICATION_REQUESTED = True
-IS_SEGMENTATION_REQUESTED = True
-IS_MOUSE_SHORTCUTS_REQUESTED = False
-IS_KEYBOARD_SHORTCUTS_REQUESTED = True
-
-MODALITY = 'CT'
-
-REQUIRE_EMPTY = True
-
-CONFIG_FILENAME = "configuration_config.yml"
-CONFIG_COPY_FILENAME = CONFIG_FILENAME.split('.')[0] + '--do-not-modify.yml'
+from utils.global_variables import *
+# From the global variables, CONFIG_FILE_PATH required the use of
+# os.path.join, which limits its usability in the utils.global.variables.py
 CONFIG_FILE_PATH = os.path.join(Path(__file__).parent.resolve(), CONFIG_FILENAME)
 
-CONF_FOLDER_NAME = '_conf'
+from scripts.SlicerCARTConfigurationInitialWindow import *
 
-CT_WINDOW_WIDTH = 90
-CT_WINDOW_LEVEL = 45
-
-
-TIMER_MUTEX = RLock()
-
-class SlicerCARTConfigurationInitialWindow(qt.QWidget):
-   def __init__(self, segmenter, parent = None):
-      super(SlicerCARTConfigurationInitialWindow, self).__init__(parent)
-
-      self.segmenter = segmenter
-
-      layout = qt.QVBoxLayout()
-
-      self.reuse_configuration_hbox = qt.QHBoxLayout()
-
-      self.new_config_radio_button = qt.QRadioButton('New configuration', self)
-      self.reuse_config_radio_button = qt.QRadioButton('Continue from existing output folder', self)
-      self.use_template_config_radio_button = qt.QRadioButton('Use template configuration', self)
-
-      self.reuse_configuration_hbox.addWidget(self.new_config_radio_button)
-      self.reuse_configuration_hbox.addWidget(self.reuse_config_radio_button)
-      self.reuse_configuration_hbox.addWidget(self.use_template_config_radio_button)
-
-      self.new_config_radio_button.toggled.connect(lambda: self.update_selected_reuse_config_option(self.new_config_radio_button.text))
-      self.reuse_config_radio_button.toggled.connect(lambda: self.update_selected_reuse_config_option(self.reuse_config_radio_button.text))
-      self.use_template_config_radio_button.toggled.connect(lambda: self.update_selected_reuse_config_option(self.use_template_config_radio_button.text))
-
-      self.new_config_radio_button.setChecked(True) # par défaut
-      self.reuse_configuration_selected_option = self.new_config_radio_button.text
-
-      layout.addLayout(self.reuse_configuration_hbox) 
-
-      self.next_button = qt.QPushButton('Next')
-      self.next_button.clicked.connect(self.push_next)
-      layout.addWidget(self.next_button)
-
-      self.cancel_button = qt.QPushButton('Cancel')
-      self.cancel_button.clicked.connect(self.push_cancel)
-      layout.addWidget(self.cancel_button)
-
-      self.setLayout(layout)
-      self.setWindowTitle("Configure SlicerCART")
-      self.resize(800, 100)
-   
-   def update_selected_reuse_config_option(self, option):
-       self.reuse_configuration_selected_option = option
-   
-   def push_next(self):
-       if self.reuse_configuration_selected_option == self.reuse_config_radio_button.text:
-           msg = qt.QMessageBox()
-           msg.setWindowTitle('Informative Message')
-           msg.setText('Please select the working output directory. ')
-           msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
-           msg.buttonClicked.connect(self.select_output_folder_clicked)
-           msg.exec()
-       elif self.reuse_configuration_selected_option == self.use_template_config_radio_button.text:
-           msg = qt.QMessageBox()
-           msg.setWindowTitle('Informative Message')
-           msg.setText('Please select the _conf folder containing the template configuration files. ')
-           msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
-           msg.buttonClicked.connect(self.select_template_folder_clicked)
-           msg.exec()
-       elif self.reuse_configuration_selected_option == self.new_config_radio_button.text:
-           slicerCARTConfigurationSetupWindow = SlicerCARTConfigurationSetupWindow(self.segmenter)
-           slicerCARTConfigurationSetupWindow.show()
-           self.segmenter.ui.SelectOutputFolder.setVisible(True)
-           self.close()
-
-   def select_output_folder_clicked(self, button):
-       if button.text == 'OK':
-          global REQUIRE_EMPTY
-          REQUIRE_EMPTY = False
-          self.segmenter.onSelectOutputFolder()
-          REQUIRE_EMPTY = True
-          self.segmenter.ui.SelectOutputFolder.setVisible(False)
-          if (os.path.exists(f'{self.segmenter.outputFolder}{os.sep}{CONF_FOLDER_NAME}') and
-             os.path.exists(f'{self.segmenter.outputFolder}{os.sep}{CONF_FOLDER_NAME}{os.sep}{CONFIG_COPY_FILENAME}')):
-                # use this configuration directly
-                shutil.copy(f'{self.segmenter.outputFolder}{os.sep}{CONF_FOLDER_NAME}{os.sep}{CONFIG_COPY_FILENAME}', CONFIG_FILE_PATH)
-                self.segmenter.setup_configuration()
-                self.close()
-          else:
-                msg = qt.QMessageBox()
-                msg.setWindowTitle('Informative Message')
-                msg.setText('The selected output folder does not contain the required configuration files for SlicerCART. Please try again. ')
-                msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
-                msg.buttonClicked.connect(self.error_msg_for_output_folder_selection_clicked)
-                msg.exec()
-          
-       else:
-          slicerCART_configuration_initial_window = SlicerCARTConfigurationInitialWindow(self.segmenter)
-          slicerCART_configuration_initial_window.show()
-          self.close()
-          return
-       
-   def error_msg_for_output_folder_selection_clicked(self, button):
-        slicerCART_configuration_initial_window = SlicerCARTConfigurationInitialWindow(self.segmenter)
-        slicerCART_configuration_initial_window.show()
-        self.close()
-   
-   def select_template_folder_clicked(self, button):
-       if button.text == 'OK':
-          conf_folder_path = qt.QFileDialog.getExistingDirectory(None,"Open a folder", '', qt.QFileDialog.ShowDirsOnly)
-          if (os.path.split(conf_folder_path)[1] == CONF_FOLDER_NAME and
-              os.path.exists(f'{conf_folder_path}{os.sep}{CONFIG_COPY_FILENAME}')):
-                 
-              slicerCARTConfigurationSetupWindow = SlicerCARTConfigurationSetupWindow(self.segmenter, conf_folder_path)
-              slicerCARTConfigurationSetupWindow.show()
-              self.segmenter.ui.SelectOutputFolder.setVisible(True)
-              self.close()
-
-          else:
-              msg = qt.QMessageBox()
-              msg.setWindowTitle('Informative Message')
-              msg.setText('The selected output folder does not contain the required configuration files for SlicerCART. Please try again. ')
-              msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
-              msg.buttonClicked.connect(self.error_msg_for_output_folder_selection_clicked)
-              msg.exec()
-              
-       else:
-          slicerCART_configuration_initial_window = SlicerCARTConfigurationInitialWindow(self.segmenter)
-          slicerCART_configuration_initial_window.show()
-          self.close()
-          return
-       
-   def push_cancel(self):
-       msg = qt.QMessageBox()
-       msg.setWindowTitle('Informative Message')
-       msg.setText('Using default configurations. To select a different configuration, restart the application. ')
-       msg.exec()
-
-       self.segmenter.setup_configuration()
-       self.close()
+#
+# class SlicerCARTConfigurationInitialWindow(qt.QWidget):
+#    def __init__(self, segmenter, parent = None):
+#       super(SlicerCARTConfigurationInitialWindow, self).__init__(parent)
+#
+#       self.segmenter = segmenter
+#
+#       layout = qt.QVBoxLayout()
+#
+#       self.reuse_configuration_hbox = qt.QHBoxLayout()
+#
+#       self.new_config_radio_button = qt.QRadioButton('New configuration', self)
+#       self.reuse_config_radio_button = qt.QRadioButton('Continue from existing output folder', self)
+#       self.use_template_config_radio_button = qt.QRadioButton('Use template configuration', self)
+#
+#       self.reuse_configuration_hbox.addWidget(self.new_config_radio_button)
+#       self.reuse_configuration_hbox.addWidget(self.reuse_config_radio_button)
+#       self.reuse_configuration_hbox.addWidget(self.use_template_config_radio_button)
+#
+#       self.new_config_radio_button.toggled.connect(lambda: self.update_selected_reuse_config_option(self.new_config_radio_button.text))
+#       self.reuse_config_radio_button.toggled.connect(lambda: self.update_selected_reuse_config_option(self.reuse_config_radio_button.text))
+#       self.use_template_config_radio_button.toggled.connect(lambda: self.update_selected_reuse_config_option(self.use_template_config_radio_button.text))
+#
+#       self.new_config_radio_button.setChecked(True) # par défaut
+#       self.reuse_configuration_selected_option = self.new_config_radio_button.text
+#
+#       layout.addLayout(self.reuse_configuration_hbox)
+#
+#       self.next_button = qt.QPushButton('Next')
+#       self.next_button.clicked.connect(self.push_next)
+#       layout.addWidget(self.next_button)
+#
+#       self.cancel_button = qt.QPushButton('Cancel')
+#       self.cancel_button.clicked.connect(self.push_cancel)
+#       layout.addWidget(self.cancel_button)
+#
+#       self.setLayout(layout)
+#       self.setWindowTitle("Configure SlicerCART")
+#       self.resize(800, 100)
+#
+#    def update_selected_reuse_config_option(self, option):
+#        self.reuse_configuration_selected_option = option
+#
+#    def push_next(self):
+#        if self.reuse_configuration_selected_option == self.reuse_config_radio_button.text:
+#            msg = qt.QMessageBox()
+#            msg.setWindowTitle('Informative Message')
+#            msg.setText('Please select the working output directory. ')
+#            msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
+#            msg.buttonClicked.connect(self.select_output_folder_clicked)
+#            msg.exec()
+#        elif self.reuse_configuration_selected_option == self.use_template_config_radio_button.text:
+#            msg = qt.QMessageBox()
+#            msg.setWindowTitle('Informative Message')
+#            msg.setText('Please select the _conf folder containing the template configuration files. ')
+#            msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
+#            msg.buttonClicked.connect(self.select_template_folder_clicked)
+#            msg.exec()
+#        elif self.reuse_configuration_selected_option == self.new_config_radio_button.text:
+#            slicerCARTConfigurationSetupWindow = SlicerCARTConfigurationSetupWindow(self.segmenter)
+#            slicerCARTConfigurationSetupWindow.show()
+#            self.segmenter.ui.SelectOutputFolder.setVisible(True)
+#            self.close()
+#
+#    def select_output_folder_clicked(self, button):
+#        if button.text == 'OK':
+#           global REQUIRE_EMPTY
+#           REQUIRE_EMPTY = False
+#           self.segmenter.onSelectOutputFolder()
+#           REQUIRE_EMPTY = True
+#           self.segmenter.ui.SelectOutputFolder.setVisible(False)
+#           if (os.path.exists(f'{self.segmenter.outputFolder}{os.sep}{CONF_FOLDER_NAME}') and
+#              os.path.exists(f'{self.segmenter.outputFolder}{os.sep}{CONF_FOLDER_NAME}{os.sep}{CONFIG_COPY_FILENAME}')):
+#                 # use this configuration directly
+#                 shutil.copy(f'{self.segmenter.outputFolder}{os.sep}{CONF_FOLDER_NAME}{os.sep}{CONFIG_COPY_FILENAME}', CONFIG_FILE_PATH)
+#                 self.segmenter.setup_configuration()
+#                 self.close()
+#           else:
+#                 msg = qt.QMessageBox()
+#                 msg.setWindowTitle('Informative Message')
+#                 msg.setText('The selected output folder does not contain the required configuration files for SlicerCART. Please try again. ')
+#                 msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
+#                 msg.buttonClicked.connect(self.error_msg_for_output_folder_selection_clicked)
+#                 msg.exec()
+#
+#        else:
+#           slicerCART_configuration_initial_window = SlicerCARTConfigurationInitialWindow(self.segmenter)
+#           slicerCART_configuration_initial_window.show()
+#           self.close()
+#           return
+#
+#    def error_msg_for_output_folder_selection_clicked(self, button):
+#         slicerCART_configuration_initial_window = SlicerCARTConfigurationInitialWindow(self.segmenter)
+#         slicerCART_configuration_initial_window.show()
+#         self.close()
+#
+#    def select_template_folder_clicked(self, button):
+#        if button.text == 'OK':
+#           conf_folder_path = qt.QFileDialog.getExistingDirectory(None,"Open a folder", '', qt.QFileDialog.ShowDirsOnly)
+#           if (os.path.split(conf_folder_path)[1] == CONF_FOLDER_NAME and
+#               os.path.exists(f'{conf_folder_path}{os.sep}{CONFIG_COPY_FILENAME}')):
+#
+#               slicerCARTConfigurationSetupWindow = SlicerCARTConfigurationSetupWindow(self.segmenter, conf_folder_path)
+#               slicerCARTConfigurationSetupWindow.show()
+#               self.segmenter.ui.SelectOutputFolder.setVisible(True)
+#               self.close()
+#
+#           else:
+#               msg = qt.QMessageBox()
+#               msg.setWindowTitle('Informative Message')
+#               msg.setText('The selected output folder does not contain the required configuration files for SlicerCART. Please try again. ')
+#               msg.setStandardButtons(qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
+#               msg.buttonClicked.connect(self.error_msg_for_output_folder_selection_clicked)
+#               msg.exec()
+#
+#        else:
+#           slicerCART_configuration_initial_window = SlicerCARTConfigurationInitialWindow(self.segmenter)
+#           slicerCART_configuration_initial_window.show()
+#           self.close()
+#           return
+#
+#    def push_cancel(self):
+#        msg = qt.QMessageBox()
+#        msg.setWindowTitle('Informative Message')
+#        msg.setText('Using default configurations. To select a different configuration, restart the application. ')
+#        msg.exec()
+#
+#        self.segmenter.setup_configuration()
+#        self.close()
 
 class SlicerCARTConfigurationSetupWindow(qt.QWidget):
    def __init__(self, segmenter, conf_folder_path = None, edit_conf = False, parent = None):
